Single file for GCP Cloud Functions
