--- src/utils.py
@@ -1,37 +1,0 @@
-from random import randint
-import requests
-import json
-
-
-def random_header() -> str:
-    """
-    Returns a random useragent header.
-    """
-    
-    headers = [
-        "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/114.0.0.0 Safari/537.36",
-        "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/114.0.0.0 Safari/537.36",
-        "Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:109.0) Gecko/20100101 Firefox/114.0",
-        "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/114.0.0.0 Safari/537.36",
-        "Mozilla/5.0 (X11; Linux x86_64; rv:109.0) Gecko/20100101 Firefox/114.0",
-        "Mozilla/5.0 (Macintosh; Intel Mac OS X 10.15; rv:109.0) Gecko/20100101 Firefox/114.0",
-        "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/113.0.0.0 Safari/537.36",
-        "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/16.5 Safari/605.1.15"
-    ]
-    
-    random_number = randint(0, len(headers) - 1)
-    
-    return headers[random_number]
-
-
-def get_mastodon_id(username: str, instance: str) -> str:
-    """
-    A quick bit of code to fetch a mastodon user's id.
-    Usage: get_mastodon_id(username='john_doe', instance='mastodon.social')
-    """
-    
-    url = f"https://{instance}/api/v1/accounts/lookup?acct={username}"
-    response = requests.get(url=url)
-    response_dict = json.loads(response.text)
-    
-    return response_dict['id']
